Optimize loading of plugin nvim-autopairs
Make lazy loading of this plugin more specific by loading it only when
bracket delimiter closing is needed(when the starting bracket delimiter
has been typed in insert mode) rather than when a file is read.
This improves performance when the `BufReadPre` event occurs because it
has one less plugin to load. Loading files is rather slow because a
majority of plugins are loaded when `BufReadPre` files.